Crate c_ares_sys

Low-level bindings for the c-ares library.

In most cases this crate should not be used directly. The c-ares crate provides a safe wrapper and should be preferred wherever possible.

Structs

Union_Unnamed1

Union_Unnamed2

Union_Unnamed3

ares_addr6ttl

ares_addr_node

ares_addr_port_node

ares_addrttl

ares_in6_addr

ares_mx_reply

ares_naptr_reply

ares_options

ares_soa_reply

ares_srv_reply

ares_txt_ext

ares_txt_reply

Enums

apattern

ares_channeldata

hostent

Constants

ARES_EADDRGETNETWORKPARAMS

ARES_EBADFAMILY

ARES_EBADFLAGS

ARES_EBADHINTS

ARES_EBADNAME

ARES_EBADQUERY

ARES_EBADRESP

ARES_EBADSTR

ARES_ECANCELLED

ARES_ECONNREFUSED

ARES_EDESTRUCTION

ARES_EFILE

ARES_EFORMERR

ARES_ELOADIPHLPAPI

ARES_ENODATA

ARES_ENOMEM

ARES_ENONAME

ARES_ENOTFOUND

ARES_ENOTIMP

ARES_ENOTINITIALIZED

ARES_EOF

ARES_EREFUSED

ARES_ESERVFAIL

ARES_ETIMEOUT

ARES_FLAG_EDNS

ARES_FLAG_IGNTC

ARES_FLAG_NOALIASES

ARES_FLAG_NOCHECKRESP

ARES_FLAG_NORECURSE

ARES_FLAG_NOSEARCH

ARES_FLAG_PRIMARY

ARES_FLAG_STAYOPEN

ARES_FLAG_USEVC

ARES_GETSOCK_MAXNUM

ARES_LIB_INIT_ALL

ARES_LIB_INIT_NONE

ARES_LIB_INIT_WIN32

ARES_NI_DCCP

ARES_NI_DGRAM

ARES_NI_IDN

ARES_NI_IDN_ALLOW_UNASSIGNED

ARES_NI_IDN_USE_STD3_ASCII_RULES

ARES_NI_LOOKUPHOST

ARES_NI_LOOKUPSERVICE

ARES_NI_NAMEREQD

ARES_NI_NOFQDN

ARES_NI_NUMERICHOST

ARES_NI_NUMERICSCOPE

ARES_NI_NUMERICSERV

ARES_NI_SCTP

ARES_NI_TCP

ARES_NI_UDP

ARES_OPT_DOMAINS

ARES_OPT_EDNSPSZ

ARES_OPT_FLAGS

ARES_OPT_LOOKUPS

ARES_OPT_NDOTS

ARES_OPT_NOROTATE

ARES_OPT_ROTATE

ARES_OPT_SERVERS

ARES_OPT_SOCK_RCVBUF

ARES_OPT_SOCK_SNDBUF

ARES_OPT_SOCK_STATE_CB

ARES_OPT_SORTLIST

ARES_OPT_TCP_PORT

ARES_OPT_TIMEOUT

ARES_OPT_TIMEOUTMS

ARES_OPT_TRIES

ARES_OPT_UDP_PORT

ARES_SOCKET_BAD

ARES_SUCCESS

Functions

ares_cancel⚠

ares_create_query⚠

ares_destroy⚠

ares_destroy_options⚠

ares_dup⚠

ares_expand_name⚠

ares_expand_string⚠

ares_fds⚠

ares_free_data⚠

ares_free_hostent⚠

ares_free_string⚠

ares_get_servers⚠

ares_get_servers_ports⚠

ares_gethostbyaddr⚠

ares_gethostbyname⚠

ares_gethostbyname_file⚠

ares_getnameinfo⚠

ares_getsock⚠

ares_inet_ntop⚠

ares_inet_pton⚠

ares_init⚠

ares_init_options⚠

ares_library_cleanup⚠

ares_library_init⚠

ares_library_init_mem⚠

ares_library_initialized⚠

ares_mkquery⚠

ares_parse_a_reply⚠

ares_parse_aaaa_reply⚠

ares_parse_mx_reply⚠

ares_parse_naptr_reply⚠

ares_parse_ns_reply⚠

ares_parse_ptr_reply⚠

ares_parse_soa_reply⚠

ares_parse_srv_reply⚠

ares_parse_txt_reply⚠

ares_parse_txt_reply_ext⚠

ares_process⚠

ares_process_fd⚠

ares_query⚠

ares_save_options⚠

ares_search⚠

ares_send⚠

ares_set_local_dev⚠

ares_set_local_ip4⚠

ares_set_local_ip6⚠

ares_set_servers⚠

ares_set_servers_csv⚠

ares_set_servers_ports⚠

ares_set_servers_ports_csv⚠

ares_set_socket_callback⚠

ares_set_socket_configure_callback⚠

ares_set_sortlist⚠

ares_strerror⚠

ares_timeout⚠

ares_version⚠

Type Definitions

__cares_rule_02__

__cares_rule_03__

ares_callback

ares_channel

ares_host_callback

ares_nameinfo_callback

ares_sock_config_callback

ares_sock_create_callback

ares_sock_state_cb

ares_socket_t

ares_socklen_t